refactor(domain): domain.participant로 위치되는 클래스 이동
