+++ src/test/java/domain/DealerTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
+import domain.participant.Dealer;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
